requirements.txt
# Databricks runtime typically includes these, but listing for local development
pyspark>=3.4.0
delta-spark>=2.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# API clients and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# AWS integration
boto3>=1.28.0
botocore>=1.31.0

# Configuration management
pyyaml>=6.0
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Databricks SDK
databricks-sdk>=0.12.0
databricks-cli>=0.18.0
